Fix: 초기화가 되어있으면 새로고침해도 다시 초기화 안하도록 수정

+++ src/domain/restaurantAPI.ts
@@ -5,7 +5,10 @@ const KEY = 'restaurant';
 
 const restaurantAPI = {
   initialize: () => {
-    window.localStorage.setItem(KEY, JSON.stringify(initialData));
+    const existingData = window.localStorage.getItem(KEY);
+    if (!existingData) {
+      window.localStorage.setItem(KEY, JSON.stringify(initialData));
+    }
   },
 
   save: (restaurant: RestaurantInfo) => {
